working on .service

+++ flask/flaskrunner.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-pkill -9 -f 'celery worker'
+if pgrep -f 'celery worker' 2>/dev/null; then
+  echo "Terminating process_name"
+  pkill -9 -f 'celery worker'
+fi
 cd /home/ubuntu
 source lysenv1/bin/activate
 cd lightyear/flask
